@@ -1,36 +1,35 @@
 {
-  "name": "singer-node",
-  "version": "2.0.0",
-  "description": "Write the singer format from node",
+  "name": "@jasonlee7903/singer-node",
+  "version": "1.0.0",
+  "description": "Write the singer format from node - Custom fork",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
     "test": "mocha",
     "build": "tsc",
     "lint": "eslint . --ext .ts",
-    "prepublishOnly": "yarn run build"
+    "prepublishOnly": "npm run build"
   },
   "files": [
     "dist/**/*"
   ],
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/Biron-BI/singer-node.git"
+    "url": "git+https://github.com/li199-code/singer-node.git"
   },
   "license": "AGPL-3.0-only",
   "bugs": {
-    "url": "https://github.com/Biron-BI/singer-node/issues"
+    "url": "https://github.com/li199-code/singer-node/issues"
   },
-  "homepage": "https://github.com/Biron-BI/singer-node#readme",
+  "homepage": "https://github.com/li199-code/singer-node#readme",
   "dependencies": {
     "@types/lodash.get": "^4.4.7",
     "@types/lodash.unset": "^4.5.7",
-    "@types/lodash.update": "^4.10.7",
     "commander": "^8.3.0",
     "json-schema-ref-parser": "^9.0.9",
+    "lodash": "^4.17.21",
     "lodash.get": "^4.4.2",
-    "lodash.unset": "^4.5.2",
-    "lodash.update": "^4.10.2"
+    "lodash.unset": "^4.5.2"
   },
   "devDependencies": {
     "@types/json-schema": "^7.0.9",
@@ -42,6 +41,6 @@
     "mocha": "^9.1.3",
     "ts-node": "^10.4.0",
     "tsconfig-paths": "^3.12.0",
-    "typescript": "^4.5.2"
+    "typescript": "^5.9.2"
   }
 }
